Rename files and add depot position generation
