+++ DataConfiguration.py
@@ -1,4 +1,4 @@
-from RandomDataGenerator import RandomDataGenerator
+from RandomDataGenerator import RandomRunnable
 from ReadFromFile import ReadFromFile
 # This file will contrain the core configuration for running the consumer and producer.
 # It will also contain the functions needed to generate data for the different topics.
