+++ Runnable.py
@@ -1,0 +1,7 @@
+import abc
+
+class Runnable(abc.ABC):
+
+    @abc.abstractmethod
+    def run(self, args):
+        pass
